refractor: conversion method in Boisson

+++ src/main/java/exo5/Quantite.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Getter
 @AllArgsConstructor
 public class Quantite {
-  private int valeur;
+  private double valeur;
   private Unite unite;
 }
